+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+ - Dashboard shows all listed tokens
+ - Token Management section now shows all listed tokens
+
+### Removed
+
+ - Token Search Bar from TokenManagement component
+
 ## [0.4.2] - 2022-03-25
 
 ### Added